Refactored the projects infrastructure.
