--- src/Chirp.Core/DataTransferObjects.cs
@@ -1,3 +1,0 @@
-namespace Chirp.Core;
-
-public record CheepDTO(string Author, string Message, long Timestamp, bool IsFollowed);
