+++ src/Chirp.Infrastructure/Cheeps/CheepRepository.cs
@@ -2,19 +2,19 @@ using Microsoft.EntityFrameworkCore;
 using Chirp.Core;
 using System.ComponentModel.DataAnnotations;
 
-using CheepDTO = Chirp.Core.CheepDTO;
+namespace Chirp.Infrastructure.Cheeps;
 
-namespace Chirp.Infrastructure;
+public record CheepDto(string Author, string Message, long Timestamp, bool IsFollowed);
 
 public interface ICheepRepository
 {
     Task AddCheep(Cheep cheep);
-    Task<List<CheepDTO>> GetCheepDTO(int page, string userName);
-    Task<List<CheepDTO>> GetCheepDTOFromAuthor(int page, string authorName, string userName);
-    Task<List<CheepDTO>> GetCheepDTOFromMe(int page, string userName);
+    Task<List<CheepDto>> GetCheepDTO(int page, string userName);
+    Task<List<CheepDto>> GetCheepDTOFromAuthor(int page, string authorName, string userName);
+    Task<List<CheepDto>> GetCheepDTOFromMe(int page, string userName);
     Task<Author> GetAuthorByName(string name);
     Task<Author> GetAuthorByEmail(string email);
-    Task<List<CheepDTO>> GetAllCheepDTOFromAuthor(string author, string userName);
+    Task<List<CheepDto>> GetAllCheepDTOFromAuthor(string author, string userName);
 }
 
 public class CheepRepository : ICheepRepository
@@ -51,49 +51,49 @@ public class CheepRepository : ICheepRepository
         return query.FirstOrDefaultAsync()!;
     }
 
-    public Task<List<CheepDTO>> GetCheepDTO(int page, string userName)
+    public Task<List<CheepDto>> GetCheepDTO(int page, string userName)
     {
         var user = _dbContext.Authors.FirstOrDefaultAsync(a => a.UserName == userName);
         var query = (from cheep in _dbContext.Cheeps
                 orderby cheep.TimeStamp descending
-                select new CheepDTO(cheep.Author.UserName ?? "", cheep.Text, (long)cheep.TimeStamp.Subtract(DateTime.UnixEpoch).TotalSeconds, user.Result != null && cheep.Author.Followers.Contains(user.Result)))
+                select new CheepDto(cheep.Author.UserName ?? "", cheep.Text, (long)cheep.TimeStamp.Subtract(DateTime.UnixEpoch).TotalSeconds, user.Result != null && cheep.Author.Followers.Contains(user.Result)))
             .Skip((page - 1) * CHEEPS_PER_PAGE)
             .Take(CHEEPS_PER_PAGE);
         return query.ToListAsync();
     }
     
-    public Task<List<CheepDTO>> GetCheepDTOFromAuthor(int page, string authorName, string userName)
+    public Task<List<CheepDto>> GetCheepDTOFromAuthor(int page, string authorName, string userName)
     {
         var user = _dbContext.Authors.FirstOrDefaultAsync(a => a.UserName == userName);
         var query = (from cheep in _dbContext.Cheeps
                 where cheep.Author.UserName == authorName
                 orderby cheep.TimeStamp descending
-                select new CheepDTO(cheep.Author.UserName ?? "", cheep.Text, (long)cheep.TimeStamp.Subtract(DateTime.UnixEpoch).TotalSeconds, user.Result != null && cheep.Author.Followers.Contains(user.Result)))
+                select new CheepDto(cheep.Author.UserName ?? "", cheep.Text, (long)cheep.TimeStamp.Subtract(DateTime.UnixEpoch).TotalSeconds, user.Result != null && cheep.Author.Followers.Contains(user.Result)))
             .Skip((page - 1) * CHEEPS_PER_PAGE)
             .Take(CHEEPS_PER_PAGE);
         return query.ToListAsync();
     }
-    public Task<List<CheepDTO>> GetCheepDTOFromMe(int page, string userName)
+    public Task<List<CheepDto>> GetCheepDTOFromMe(int page, string userName)
     {
         var user = _dbContext.Authors.FirstOrDefaultAsync(a => a.UserName == userName);
         var query = (from cheep in _dbContext.Cheeps
                 where cheep.Author.UserName == userName || 
                       cheep.Author.Followers.Any(f => f.UserName == userName)
                 orderby cheep.TimeStamp descending
-                select new CheepDTO(cheep.Author.UserName ?? "", cheep.Text, (long)cheep.TimeStamp.Subtract(DateTime.UnixEpoch).TotalSeconds, user.Result != null && cheep.Author.Followers.Contains(user.Result)))
+                select new CheepDto(cheep.Author.UserName ?? "", cheep.Text, (long)cheep.TimeStamp.Subtract(DateTime.UnixEpoch).TotalSeconds, user.Result != null && cheep.Author.Followers.Contains(user.Result)))
             .Skip((page - 1) * CHEEPS_PER_PAGE)
             .Take(CHEEPS_PER_PAGE);
         return query.ToListAsync();
     }
     
     // Removed cheep.Author.Followers.Any(f => f.UserName == author) from the query and the .Skip(0) and .Take(CHEEPS_PER_PAGE); as no pagination is needed is this method.
-    public Task<List<CheepDTO>> GetAllCheepDTOFromAuthor(string author, string userName)
+    public Task<List<CheepDto>> GetAllCheepDTOFromAuthor(string author, string userName)
     {
         var user = _dbContext.Authors.FirstOrDefaultAsync(a => a.UserName == userName);
         var query = from cheep in _dbContext.Cheeps
             where cheep.Author.UserName == author
             orderby cheep.TimeStamp ascending
-            select new CheepDTO(
+            select new CheepDto(
                 cheep.Author.UserName ?? "", 
                 cheep.Text, 
                 (long)cheep.TimeStamp.Subtract(DateTime.UnixEpoch).TotalSeconds,
