+++ src/Chirp.Web/Areas/Identity/Pages/Account/Register.cshtml.cs
@@ -19,7 +19,8 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.AspNetCore.WebUtilities;
 using Microsoft.Extensions.Logging;
-using Chirp.Infrastructure;
+using Chirp.Infrastructure.Cheeps;
+using Chirp.Infrastructure.Authors;
 
 namespace Chirp.Web.Areas.Identity.Pages.Account
 {
