+++ src/Chirp.Web/Pages/Public.cshtml.cs
@@ -2,7 +2,8 @@
 
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using Chirp.Infrastructure;
+using Chirp.Infrastructure.Cheeps;
+using Chirp.Infrastructure.Authors;
 using Chirp.Web.Pages.Shared.Models;
 using Microsoft.AspNetCore.Identity;
 using Chirp.Core;
