+++ src/Chirp.Web/Startup.cs
@@ -13,6 +13,8 @@
 
 using Chirp.Core;
 using Chirp.Infrastructure;
+using Chirp.Infrastructure.Cheeps;
+using Chirp.Infrastructure.Authors;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Data.Sqlite;
